Attempt to guard driver's synchronization context against caller's misusage is troubled. In AspNet sync context, there are many limitations when callback is called.
If driver calls something when AspNet does not expects it, exception will be thrown in result.

Resolution is to expose OnMessageArrived in driver's synchronization context, and document requirement to not to delay message handling. I can't find any way to enforce "ObserveOn must be called for given Rx pipeline" rule.